Implemented detecing browser language Put auth guard on index route Fixed no delete confirmation for training

+++ src/app/trainings/unit-leader/training-management/training-management.component.ts
@@ -26,8 +26,10 @@ export class TrainingManagementComponent implements OnInit {
     }
 
     deleteTraining(): void {
-        this.trainingsService.delete(this.training).subscribe(
-            () => this.router.navigate(['/unitleader/trainings'])
-        );
+        if (confirm(this.deleteDialogMessage)) {
+            this.trainingsService.delete(this.training).subscribe(
+                () => this.router.navigate(['/unitleader/trainings'])
+            );
+        }
     }
 }
